fix: proto files format and cleanup
