+++ proto/sge/bet/bet_odds.proto
@@ -29,9 +29,9 @@ message BetOdds {
 
   // max_loss_multiplier is the factor for calculating max loss for given odds
   string max_loss_multiplier = 4 [
-    (gogoproto.customname) = "MaxLossMultiplier" ,
+    (gogoproto.customname) = "MaxLossMultiplier",
     (gogoproto.jsontag) = "max_loss_multiplier",
-     json_name = "max_loss_multiplier",
+    json_name = "max_loss_multiplier",
     (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Dec",
     (gogoproto.nullable) = false
   ];
