+++ proto/sge/bet/kyc.proto
@@ -10,7 +10,7 @@ message KycDataPayload {
   // kyc_approved represent if kvc validation is approved or not.
   bool kyc_approved = 1;
   // kyc_id is the id of the kyc.
-  string kyc_id = 2[
+  string kyc_id = 2 [
     (gogoproto.customname) = "KycID",
     (gogoproto.jsontag) = "kyc_id",
     json_name = "kyc_id"
