+++ proto/sge/house/house.proto
@@ -5,112 +5,97 @@ import "gogoproto/gogo.proto";
 
 option go_package = "github.com/sge-network/sge/x/house/types";
 
-// Params defines the parameters for the house module.
-message Params {
-  option (gogoproto.equal)            = true;
-  option (gogoproto.goproto_stringer) = false;
-
-  // minimum_deposit is the minum amount of deposit acceptable.
-  string minimum_deposit = 1 [
-    (gogoproto.moretags) = "yaml:\"minimum_deposit\"",
-    (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Int",
-    (gogoproto.nullable) = false
-  ];
-
-  // house_participation_fee is the % of deposit to be paid for house participation by the user
-  string house_participation_fee = 2 [
-    (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Dec",
-    (gogoproto.nullable) = false
-  ];
-}
-
 // Deposit represents the deposit against a sport event held by an account.
 message Deposit {
-    option (gogoproto.equal)            = false;
-    option (gogoproto.goproto_getters)  = false;
-    option (gogoproto.goproto_stringer) = false;
+  option (gogoproto.equal) = false;
+  option (gogoproto.goproto_getters) = false;
+  option (gogoproto.goproto_stringer) = false;
 
   // creator is the bech32-encoded address of the depositor.
-  string creator = 1 [(gogoproto.moretags) = "yaml:\"creator\""];
+  string creator = 1 [ (gogoproto.moretags) = "yaml:\"creator\"" ];
 
-  // sport_event_uid is the uid of sport event against which deposit is being made.
+  // sport_event_uid is the uid of sport event against which deposit is being
+  // made.
   string sport_event_uid = 2 [
     (gogoproto.customname) = "SportEventUID",
     (gogoproto.jsontag) = "sport_event_uid",
     json_name = "sport_event_uid"
   ];
 
+  // participation_index index corresponding to the book participation
+  uint64 participation_index = 3
+      [ (gogoproto.moretags) = "yaml:\"participation_index\"" ];
+
   // amount is the amount being deposited.
-  string amount = 3 [
+  string amount = 4 [
     (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Int",
-    (gogoproto.nullable)   = false,
-    (gogoproto.moretags)   = "yaml:\"amount\""
+    (gogoproto.nullable) = false,
+    (gogoproto.moretags) = "yaml:\"amount\""
   ];
 
-  // fee is deducted from amount at the point of deposit.
-  string fee = 4 [
+  // fee is deducted from the amount on deposition.
+  string fee = 5 [
     (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Int",
-    (gogoproto.nullable)   = false,
-    (gogoproto.moretags)   = "yaml:\"fee\""
+    (gogoproto.nullable) = false,
+    (gogoproto.moretags) = "yaml:\"fee\""
   ];
 
   // liquidity is the liquidity being provided to the house after fee deduction.
-  string liquidity = 5 [
+  string liquidity = 6 [
     (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Int",
-    (gogoproto.nullable)   = false,
-    (gogoproto.moretags)   = "yaml:\"liquidity\""
+    (gogoproto.nullable) = false,
+    (gogoproto.moretags) = "yaml:\"liquidity\""
   ];
 
-  // participation_index index corresponding to the book participation
-  uint64 participation_index = 6 [(gogoproto.moretags) = "yaml:\"participation_index\""];
-
   // withdrawal_count is the total count of the withdrawal attempts
-  uint64 withdrawal_count = 7 [(gogoproto.moretags) = "yaml:\"withdrawals\""];
+  uint64 withdrawal_count = 7 [ (gogoproto.moretags) = "yaml:\"withdrawals\"" ];
 
   // total_withdrawal_amount is the total amount of withdrawal attempts
   string total_withdrawal_amount = 8 [
     (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Int",
-    (gogoproto.nullable)   = false,
-    (gogoproto.moretags)   = "yaml:\"total_withdrawal_amount\""
+    (gogoproto.nullable) = false,
+    (gogoproto.moretags) = "yaml:\"total_withdrawal_amount\""
   ];
 }
 
 // Withdrawal represents the withdrawal against a deposit.
 message Withdrawal {
-  option (gogoproto.equal)            = false;
-  option (gogoproto.goproto_getters)  = false;
+  option (gogoproto.equal) = false;
+  option (gogoproto.goproto_getters) = false;
   option (gogoproto.goproto_stringer) = false;
 
-// withdrawal number is the number of withdrawal against the deposit
-uint64 id = 1 [
-  (gogoproto.customname) = "ID",
-  (gogoproto.jsontag) = "id",
-  json_name = "id",
-  (gogoproto.moretags) = "yaml:\"id\""
-];
+  // withdrawal is the withdrawal attempt id.
+  uint64 id = 1 [
+    (gogoproto.customname) = "ID",
+    (gogoproto.jsontag) = "id",
+    json_name = "id",
+    (gogoproto.moretags) = "yaml:\"id\""
+  ];
 
-// creator is the bech32-encoded address of the depositor.
-string creator = 2 [(gogoproto.moretags) = "yaml:\"creator\""];
+  // creator is the bech32-encoded address of the depositor.
+  string creator = 2 [ (gogoproto.moretags) = "yaml:\"creator\"" ];
 
-// sport_event_uid is the uid of sport event against which the deposit is being made.
-string sport_event_uid = 3 [
-  (gogoproto.customname) = "SportEventUID",
-  (gogoproto.jsontag) = "sport_event_uid",
-  json_name = "sport_event_uid"
-];
+  // sport_event_uid is the uid of sport-event against which the deposit is
+  // being made.
+  string sport_event_uid = 3 [
+    (gogoproto.customname) = "SportEventUID",
+    (gogoproto.jsontag) = "sport_event_uid",
+    json_name = "sport_event_uid"
+  ];
 
-// participation_index is the id corresponding to the book participation
-uint64 participation_index = 4 [(gogoproto.moretags) = "yaml:\"participation_index\""];
+  // participation_index is the id corresponding to the book participation
+  uint64 participation_index = 4
+      [ (gogoproto.moretags) = "yaml:\"participation_index\"" ];
 
-// mode is the withdrawal mode enum value
-WithdrawalMode mode = 5 [(gogoproto.moretags) = "yaml:\"mode\""];
+  // mode is the withdrawal mode enum value
+  WithdrawalMode mode = 5 [ (gogoproto.moretags) = "yaml:\"mode\"" ];
 
-// amount is the amount being withdrawan.
-string amount = 6 [
-  (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Int",
-  (gogoproto.nullable)   = false,
-  (gogoproto.moretags)   = "yaml:\"amount\""
-];
+  // amount is the amount being withdrawn.
+  string amount = 6 [
+    (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Int",
+    (gogoproto.nullable) = false,
+    (gogoproto.moretags) = "yaml:\"amount\""
+  ];
 }
 
 // WithdrawalMode is the enum type for the withdrawal mode.
