+++ proto/sge/house/tx.proto
@@ -8,25 +8,28 @@ option go_package = "github.com/sge-network/sge/x/house/types";
 
 // Msg defines the house Msg service.
 service Msg {
-  // Deposit defines a method for performing a deposit of coins to become part of the house corresponding to a sport event.
+  // Deposit defines a method for performing a deposit of coins to become part
+  // of the house corresponding to a sport event.
   rpc Deposit(MsgDeposit) returns (MsgDepositResponse);
 
-  // Withdraw defines a method for performing a withdrawal of coins of unused amount corresponding to a deposit.
+  // Withdraw defines a method for performing a withdrawal of coins of unused
+  // amount corresponding to a deposit.
   rpc Withdraw(MsgWithdraw) returns (MsgWithdrawResponse);
 }
 
-// MsgDeposit defines a SDK message for performing a deposit of coins to become part of the house corresponding to a sport event.
+// MsgDeposit defines a SDK message for performing a deposit of coins to become
+// part of the house corresponding to a sport event.
 message MsgDeposit {
-  option (gogoproto.equal)           = false;
+  option (gogoproto.equal) = false;
   option (gogoproto.goproto_getters) = false;
 
-  string                   creator = 1 [(gogoproto.moretags) = "yaml:\"creator\""];
-  string                   sport_event_uid = 2 [
+  string creator = 1 [ (gogoproto.moretags) = "yaml:\"creator\"" ];
+  string sport_event_uid = 2 [
     (gogoproto.customname) = "SportEventUID",
     (gogoproto.jsontag) = "sport_event_uid",
     json_name = "sport_event_uid"
   ];
-  string amount            = 3 [
+  string amount = 3 [
     (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Int",
     (gogoproto.nullable) = false
   ];
@@ -40,23 +43,26 @@ message MsgDepositResponse {
     json_name = "sport_event_uid"
   ];
 
-  uint64 participation_index = 2 [(gogoproto.moretags) = "yaml:\"participation_index\""];
+  uint64 participation_index = 2
+      [ (gogoproto.moretags) = "yaml:\"participation_index\"" ];
 }
 
-// MsgWithdraw defines a SDK message for performing a withdrawal of coins of unused amount corresponding to a deposit.
+// MsgWithdraw defines a SDK message for performing a withdrawal of coins of
+// unused amount corresponding to a deposit.
 message MsgWithdraw {
-  option (gogoproto.equal)           = false;
+  option (gogoproto.equal) = false;
   option (gogoproto.goproto_getters) = false;
 
-  string creator = 1 [(gogoproto.moretags) = "yaml:\"creator\""];
+  string creator = 1 [ (gogoproto.moretags) = "yaml:\"creator\"" ];
   string sport_event_uid = 2 [
     (gogoproto.customname) = "SportEventUID",
     (gogoproto.jsontag) = "sport_event_uid",
     json_name = "sport_event_uid"
   ];
-  uint64 participation_index = 3 [(gogoproto.moretags) = "yaml:\"participation_index\""];
-  WithdrawalMode mode = 4 [(gogoproto.moretags) = "yaml:\"mode\""];
-  string amount            = 5 [
+  uint64 participation_index = 3
+      [ (gogoproto.moretags) = "yaml:\"participation_index\"" ];
+  WithdrawalMode mode = 4 [ (gogoproto.moretags) = "yaml:\"mode\"" ];
+  string amount = 5 [
     (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Int",
     (gogoproto.nullable) = false
   ];
@@ -77,5 +83,6 @@ message MsgWithdrawResponse {
     json_name = "sport_event_uid"
   ];
 
-  uint64 participation_index = 3 [(gogoproto.moretags) = "yaml:\"participation_index\""];
+  uint64 participation_index = 3
+      [ (gogoproto.moretags) = "yaml:\"participation_index\"" ];
 }
