+++ proto/sge/orderbook/exposure.proto
@@ -1,0 +1,75 @@
+syntax = "proto3";
+package sgenetwork.sge.orderbook;
+
+import "gogoproto/gogo.proto";
+
+option go_package = "github.com/sge-network/sge/x/orderbook/types";
+
+// BookOddsExposure represents the exposures taken on odds
+message BookOddsExposure {
+  option (gogoproto.equal) = false;
+  option (gogoproto.goproto_getters) = false;
+  option (gogoproto.goproto_stringer) = false;
+
+  // book_id is id corresponding to the book
+  string book_id = 1 [
+    (gogoproto.customname) = "BookID",
+    (gogoproto.jsontag) = "book_id",
+    json_name = "book_id"
+  ];
+
+  // odds_id is odd'd uid
+  string odds_id = 2 [
+    (gogoproto.customname) = "OddsID",
+    (gogoproto.jsontag) = "odds_id",
+    json_name = "odds_id"
+  ];
+
+  repeated uint64 fulfillment_queue = 3
+      [ (gogoproto.moretags) = "yaml:\"fulfillment_queue\"" ];
+}
+
+// ParticipationExposure represents the exposures taken on odds by
+// participations
+message ParticipationExposure {
+  option (gogoproto.equal) = false;
+  option (gogoproto.goproto_getters) = false;
+  option (gogoproto.goproto_stringer) = false;
+
+  // book_id is id corresponding to the book
+  string book_id = 1 [
+    (gogoproto.customname) = "BookID",
+    (gogoproto.jsontag) = "book_id",
+    json_name = "book_id"
+  ];
+
+  // odds_id is odd's uid
+  string odds_id = 2 [
+    (gogoproto.customname) = "OddsID",
+    (gogoproto.jsontag) = "odds_id",
+    json_name = "odds_id"
+  ];
+
+  // participation_index is the id of initial participation queue
+  uint64 participation_index = 3
+      [ (gogoproto.moretags) = "yaml:\"participation_index\"" ];
+
+  // exposure is the total exposure taken on given odd
+  string exposure = 4 [
+    (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Int",
+    (gogoproto.nullable) = false,
+    (gogoproto.moretags) = "yaml:\"exposure\""
+  ];
+
+  // bet_amount is the total bet amount corresponding to the exposure
+  string bet_amount = 5 [
+    (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Int",
+    (gogoproto.nullable) = false,
+    (gogoproto.moretags) = "yaml:\"bet_amount\""
+  ];
+
+  bool is_fulfilled = 6 [ (gogoproto.moretags) = "yaml:\"is_fulfilled\"" ];
+
+  // number of current round in queue
+  uint64 round = 7 [ (gogoproto.moretags) = "yaml:\"round\"" ];
+}
