+++ proto/sge/orderbook/genesis.proto
@@ -3,28 +3,33 @@ package sgenetwork.sge.orderbook;
 
 import "gogoproto/gogo.proto";
 import "sge/orderbook/params.proto";
-import "sge/orderbook/book.proto";
+import "sge/orderbook/orderbook.proto";
 import "sge/orderbook/participation.proto";
 import "sge/orderbook/stats.proto";
+import "sge/orderbook/exposure.proto";
 
 option go_package = "github.com/sge-network/sge/x/orderbook/types";
 
 // GenesisState defines the orderbook module's genesis state.
 message GenesisState {
   // params defines all the parameters related to orderbook.
-  Params params = 1 [(gogoproto.nullable) = false];
+  Params params = 1 [ (gogoproto.nullable) = false ];
 
   // book_list defines the books active at genesis.
-  repeated OrderBook book_list = 2 [(gogoproto.nullable) = false];
+  repeated OrderBook book_list = 2 [ (gogoproto.nullable) = false ];
 
   // book_participation_list defines the book participations active at genesis.
-  repeated BookParticipation book_participation_list = 3 [(gogoproto.nullable) = false];
+  repeated BookParticipation book_participation_list = 3
+      [ (gogoproto.nullable) = false ];
 
   // book_exposure_list defines the book exposures active at genesis.
-  repeated BookOddsExposure book_exposure_list = 4 [(gogoproto.nullable) = false];
+  repeated BookOddsExposure book_exposure_list = 4
+      [ (gogoproto.nullable) = false ];
 
-  // participation_exposure_list defines the participation exposures active at genesis.
-  repeated ParticipationExposure participation_exposure_list = 5 [(gogoproto.nullable) = false];
+  // participation_exposure_list defines the participation exposures active at
+  // genesis.
+  repeated ParticipationExposure participation_exposure_list = 5
+      [ (gogoproto.nullable) = false ];
 
   // stats is the statistics of the order-book
   OrderBookStats stats = 6 [ (gogoproto.nullable) = false ];
