+++ proto/sge/orderbook/params.proto
@@ -7,12 +7,14 @@ option go_package = "github.com/sge-network/sge/x/orderbook/types";
 
 // Params defines the parameters for the orderbook module.
 message Params {
-    option (gogoproto.equal)            = true;
-    option (gogoproto.goproto_stringer) = false;
-  
-    // max_book_participations is the maximum number of participations per book.
-    uint64 max_book_participations = 1 [(gogoproto.moretags) = "yaml:\"max_book_participations\""];
-  
-    // batch_settlement_count is the batch settlement deposit counts.
-    uint64 batch_settlement_count = 2 [(gogoproto.moretags) = "yaml:\"batch_settlement_count\""];
-  }
+  option (gogoproto.equal) = true;
+  option (gogoproto.goproto_stringer) = false;
+
+  // max_book_participations is the maximum number of participations per book.
+  uint64 max_book_participations = 1
+      [ (gogoproto.moretags) = "yaml:\"max_book_participations\"" ];
+
+  // batch_settlement_count is the batch settlement deposit counts.
+  uint64 batch_settlement_count = 2
+      [ (gogoproto.moretags) = "yaml:\"batch_settlement_count\"" ];
+}
