+++ proto/sge/orderbook/stats.proto
@@ -5,7 +5,6 @@ option go_package = "github.com/sge-network/sge/x/orderbook/types";
 
 // OrderBookStats holds statistics of the order-book
 message OrderBookStats {
-    // resolved_unsettled is the list of book ids that needs to be settled.
-    repeated string resolved_unsettled = 1;
-  }
-  
+  // resolved_unsettled is the list of book ids that needs to be settled.
+  repeated string resolved_unsettled = 1;
+}
