+++ proto/sge/sportevent/sport_event.proto
@@ -51,18 +51,19 @@ message SportEvent {
   bool active = 10;
   // meta contains human-readable metadata of the sport-event.
   string meta = 11;
-  // sr_contribution_for_house is the amount of contibution of house in the sr 
-  string sr_contribution_for_house = 12[
-    (gogoproto.customname) = "SrContributionForHouse" ,
+  // sr_contribution_for_house is the amount of contibution of house in the sr
+  string sr_contribution_for_house = 12 [
+    (gogoproto.customname) = "SrContributionForHouse",
     (gogoproto.jsontag) = "sr_contribution_for_house",
     json_name = "sr_contribution_for_house",
     (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Int",
-    (gogoproto.nullable)   = false];
+    (gogoproto.nullable) = false
+  ];
   // book_id is the id of book
   string book_id = 13 [
-    (gogoproto.customname) = "BookID" ,
+    (gogoproto.customname) = "BookID",
     (gogoproto.jsontag) = "book_id",
-     json_name = "book_id"
+    json_name = "book_id"
   ];
 }
