+++ proto/sge/sportevent/ticket.proto
@@ -52,10 +52,13 @@ message SportEventAddTicketPayload {
   bool active = 10;
   // meta contains human-readable metadata of the sport-event.
   string meta = 11;
-  string sr_contribution_for_house = 12[
-    (gogoproto.customname) = "SrContributionForHouse" ,(gogoproto.jsontag) = "sr_contribution_for_house", json_name = "sr_contribution_for_house",
+  string sr_contribution_for_house = 12 [
+    (gogoproto.customname) = "SrContributionForHouse",
+    (gogoproto.jsontag) = "sr_contribution_for_house",
+    json_name = "sr_contribution_for_house",
     (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Int",
-    (gogoproto.nullable)   = false];
+    (gogoproto.nullable) = false
+  ];
 }
 
 // SportEventUpdateTicketPayload indicates data of update sport-event ticket
