+++ x/orderbook/types/exposure.go
@@ -5,6 +5,23 @@ import (
 	yaml "gopkg.in/yaml.v2"
 )
 
+// NewBookOddsExposure creates a new book odds exposure object
+//
+//nolint:interfacer
+func NewBookOddsExposure(bookID, oddsID string, fulfillmentQueue []uint64) BookOddsExposure {
+	return BookOddsExposure{
+		BookID:           bookID,
+		OddsID:           oddsID,
+		FulfillmentQueue: fulfillmentQueue,
+	}
+}
+
+// String returns a human readable string representation of a BookOddsExposure.
+func (boe BookOddsExposure) String() string {
+	out, _ := yaml.Marshal(boe)
+	return string(out)
+}
+
 // NewParticipationExposure creates a new participation exposure object
 //
 //nolint:interfacer
