+++ x/orderbook/types/orderbook.go
@@ -5,10 +5,10 @@ import (
 	yaml "gopkg.in/yaml.v2"
 )
 
-// NewBook creates a new orderbook object
+// NewOrderBook creates a new orderbook object
 //
 //nolint:interfacer
-func NewBook(bookID string, participationCount, oddsCount uint64, status OrderBookStatus) OrderBook {
+func NewOrderBook(bookID string, participationCount, oddsCount uint64, status OrderBookStatus) OrderBook {
 	return OrderBook{
 		ID:                 bookID,
 		ParticipationCount: participationCount,
